Split merged chat request tests
